requirements.txt
aiohttp==3.9.5
bidict==0.23.1
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-SocketIO==5.3.6
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
python-engineio==4.9.0
python-socketio==5.11.2
simple-websocket==1.0.0
Werkzeug==3.0.3
wsproto==1.2.0
PyAudio==0.2.14
pyttsx3==2.90
SpeechRecognition==3.10.4
setuptools==69.5.1
ollama==0.2.0
